Add support for wicked2nm in container workflow
In the container workflow no network needs to be configured
because host networking is used. However, for network technology
migrations like wicked to NetworkManager we have to run a
slightly modified setup_host_network service. Compared to the
standard service the container based network service does not
activate the nm connections, because host networking is active
and also stops the NetworkManager in the container instance
to avoid potential conflicts with the network manager tooling
running on the host e.g. wicked