Fix order of reboot service
Make sure the reboot service is started after the potential
apparmor-selinux service. Without the ordering the reboot
service starts while the apparmor-selinux is installing
additional software. Please note: The apparmor-selinux is
intentionally not set as a required service because depending
on the migration target apparmor-selinux is enabled or not.
For example SLE12-to-SLE15 migrations does not enable
apparmor-selinux migration.