Refactor mount_system service
So far the mount_system unit run through all disk devices
and uses the first one providing a system with an fstab file.
This concept is weak and fails on machines with multiple
linux OS disks attached. This commit refactors mount_system
such that it reads the migration_target= variable from the
kernel cmdline. The kernel cmdline gets this information
added by either the grub.d/99_migration menuentry or via
the run_migration kexec method. The migration_target=
variable is expected to contain the rootfs UUID value
of the host that should be migrated. This is a unique
identifier and as such cannot be confused by other devices
attached to the system.