Add DropComponents class
Implement DropComponents class to allow for removal of
packages, files and directories. Every service class that
has the need to perform this task needs to inherit from
DropComponents. Along with the new class also all service
units are turned into classes for consistency and to allow
the inheritance model for eventually further extensions
to the service units. This is related to Issue #421