Improve dracut logging
If no action is needed, a message should be part of the log
In addition fix the misleading systemd unit title for this
service. The regenerate initrd service runs to create a
generic (no hostonly) initrd due to a configuration request.
This has nothing to do with multipath in the first place.