Skip migration if wicked is not the default network config service
While wicked was the default network configuration code in SLE 15,
NetworkManager was also available either via the Desktop Module or
PackageHub. As such it is conceivable that a user does not have wicked
on the system to be migrated. Check if wicked is setup to manage the
network before attempting the migration to NetworkManager.