Update setup_host_network service
So far the service to clone the network setup from the host into
the live migration system was only able to replicate the standard
network setup from data in /etc/sysconfig/network. This patch
makes the service to also replicate NetworkManager connections
if NetworkManager is the active networking setup on the host
to migrate. In addition the wicked-to-networkmanager migration
only runs if wicked is the active networking setup on the
host to migrate.

This covers the following use cases:

1. SLE15 to SLE16 migration if the (sle15)host to migrate is NetworkManager based
2. SLE16 to SLE_next_generation where NetworkManager is the default

This does NOT cover the SLE12 to SLE15 migration if the (sle12)host
is NetworkManager based. That's because the SLE15 migration
live images are not NetworkManager based and cannot setup the
network even with the bind mounted etc/NetworkManager data.