Fix mount_system unit
The mount_system unit unconditionally bind mounts /run locations
into the system to migrate. If the migration system doesn't
provide these mountpoints e.g /run/NetworkManager on a SLE12-to-15
migration the mount_system unit fails and the entire migration
fails because of this mandatory unit to fail. We really have to
be more careful when writing DMS code.